refacto: moved usage to usage.py file
